+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 Focusses on making what benchee print out configurable to make it fit to your preferences
 
 ## Features (User Facing)
-* the configuration now has a `:print` key where it is possible to configure in a map what benchee prints out during benchmarking. All options are enabled by default (true). Options are:
+* The configuration now has a `:print` key where it is possible to configure in a map what benchee prints out during benchmarking. All options are enabled by default (true). Options are:
   * `:benchmarking`  - print when Benchee starts benchmarking a new job (Benchmarking name ..)
   * `:comparison` - the comparison of the different benchmarking jobs (x times slower than)
   * `:configuration` - a summary of configured benchmarking options including estimated total run time is printed before benchmarking starts
   * `:fast_warning` - warnings are displayed if functions are executed too    fast leading to inaccurate measures
+* The pre benchmarking configuration now also prints the currently used Erlang and Elixir versions (similar to `elixir -v`) 
 
 # 0.3.0 (July 11, 2016)